# strings: Optimized ToTitle, tests and some cleanup