… strings/helper.go