### Description:
I went through the strings package and did some cleanup in the tests.
Also, I optimised the "ToTitle" function - here's the benchstat:

#### ⏱ `sec/op`
| Exceptions | Before     | After      | Δ vs. base |
|------------|------------|------------|------------|
| nil        | 518.2n ± 2% | 394.8 ±1% | **-23.81%** |
| 0          | 526.4n ± 1% | 393.5n ± 1% | **-25.25%** |
| 1          | 573.1n ± 3% | 476.2n ± 0% | **-16.91%** |
| 2          | 556.0n ± 1% | 478.5n ± 1% | **-13.94%** |
| **geomean**| **543.0n**  | **433.8n** | **-20.11%** |

#### 📦 `B/op`
| Exceptions | Before     | After     | Δ vs. base |
|------------|------------|-----------|------------|
| nil        | 384.0 ± 0% | 296.0 ± 0%| **-22.92%** |
| 0          | 384.0 ± 0% | 296.0 ± 0%| **-22.92%** |
| 1          | 384.0 ± 0% | 296.0 ± 0%| **-22.92%** |
| 2          | 376.0 ± 0% | 296.0 ± 0%| **-21.28%** |
| **geomean**| **382.0**  | **296.0** | **-22.51%** |

#### 🗂 `allocs/op`
| Exceptions | Before     | After     | Δ vs. base |
|------------|------------|-----------|------------|
| nil        | 17.000 ± 0%| 3.000 ± 0%| **-82.35%** |
| 0          | 17.000 ± 0%| 3.000 ± 0%| **-82.35%** |
| 1          | 15.000 ± 0%| 3.000 ± 0%| **-80.00%** |
| 2          | 14.000 ± 0%| 3.000 ± 0%| **-78.57%** |
| **geomean**| **15.70**  | **3.000** | **-80.89%** |

### Checklist:  
- [x] **Tests Passing**: Verify by running `make test`.  
- [x] **Golint Passing**: Confirm by running `make lint`.  

#### If added a new utility function or updated existing function logic:
* [ ] Updated the utility package `EXAMPLES.md`
* [ ] Updated the utility package `README.md`

#### If added a new utility package:
* [ ] Updated the main `README.md` utility packages table


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Streamlined and optimized string utilities for better Unicode correctness, memory use, and performance without changing public APIs.
  * Removed an internal helper and consolidated logic to reduce intermediate allocations and simplify control flow.

* **Tests**
  * Added extensive new tests (including Unicode/emoji and edge cases), improved assertions, parallelized many subtests, and expanded benchmarks for clearer measurements.

* **Chores**
  * Minor cleanup and test/benchmark hygiene improvements.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->